Index trait, EdgeIndex and HexagonIndex
