+++ h3ron/src/experimental.rs
@@ -3,8 +3,8 @@ use std::result::Result;
 use h3ron_h3_sys::H3Index;
 
 use crate::error::Error;
-use crate::index::Index;
-use crate::HasH3Index;
+use crate::hexagon_index::HexagonIndex;
+use crate::Index;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct CoordIJ {
@@ -18,7 +18,7 @@ impl Default for CoordIJ {
     }
 }
 
-pub fn h3_to_local_ij(origin_index: &Index, index: &Index) -> Result<CoordIJ, Error> {
+pub fn h3_to_local_ij(origin_index: &HexagonIndex, index: &HexagonIndex) -> Result<CoordIJ, Error> {
     unsafe {
         let mut cij = h3ron_h3_sys::CoordIJ { i: 0, j: 0 };
         if h3ron_h3_sys::experimentalH3ToLocalIj(origin_index.h3index(), index.h3index(), &mut cij)
@@ -31,7 +31,10 @@ pub fn h3_to_local_ij(origin_index: &Index, index: &Index) -> Result<CoordIJ, Er
     }
 }
 
-pub fn local_ij_to_h3(origin_index: &Index, coordij: &CoordIJ) -> Result<Index, Error> {
+pub fn local_ij_to_h3(
+    origin_index: &HexagonIndex,
+    coordij: &CoordIJ,
+) -> Result<HexagonIndex, Error> {
     unsafe {
         let cij = h3ron_h3_sys::CoordIJ {
             i: coordij.i,
@@ -41,7 +44,7 @@ pub fn local_ij_to_h3(origin_index: &Index, coordij: &CoordIJ) -> Result<Index, 
         if h3ron_h3_sys::experimentalLocalIjToH3(origin_index.h3index(), &cij, &mut h3_index_out)
             == 0
         {
-            Ok(Index::new(h3_index_out))
+            Ok(HexagonIndex::new(h3_index_out))
         } else {
             Err(Error::NoLocalIJCoordinates)
         }
@@ -51,12 +54,12 @@ pub fn local_ij_to_h3(origin_index: &Index, coordij: &CoordIJ) -> Result<Index, 
 #[cfg(test)]
 mod tests {
     use crate::experimental::{h3_to_local_ij, local_ij_to_h3};
-    use crate::index::Index;
+    use crate::hexagon_index::HexagonIndex;
     use std::convert::TryFrom;
 
     #[test]
     fn test_local_ij() {
-        let origin_index = Index::try_from(0x89283080ddbffff_u64).unwrap();
+        let origin_index = HexagonIndex::try_from(0x89283080ddbffff_u64).unwrap();
         let ring = origin_index.k_ring(1);
         assert_ne!(ring.len(), 0);
         let other_index = ring.iter().find(|i| **i != origin_index).unwrap().clone();
